+++ whatsNew.md
@@ -1,5 +1,7 @@
+ * Leadership Changes during each Season plotly/CT Dec 15 2015
+ * % Player Games played 90 Minutes plotly/CT -  Dec 14 2015
+ * Scoreline heatmap with crosstalk table - Dec 10, 2015
  * Games Since min goals for/against - Dec 3, 2015
- * year on Year Changes Points/Goals etc. - Nov 30, 2015
  * Played for two clubs - Nov 28, 2015
  * Player Scoring Sequences (plotly version) - Nov 21, 2015
  * Player Scoring Sequences - November 21, 2015